backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
bcrypt==4.0.1
Pillow==10.0.0
python-dotenv==1.0.0
geopy==2.4.0
gunicorn==21.2.0
requests==2.31.0